\name{testRcppArmadillo-package}
\alias{testRcppArmadillo-package}
\alias{testRcppArmadillo}
\docType{package}
\title{
\packageTitle{testRcppArmadillo}
}
\description{
\packageDescription{testRcppArmadillo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testRcppArmadillo}
\packageIndices{testRcppArmadillo}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{testRcppArmadillo}

Maintainer: \packageMaintainer{testRcppArmadillo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
