\name{TestRcppArmadillo-package}
\alias{TestRcppArmadillo-package}
\alias{TestRcppArmadillo}
\docType{package}
\title{
\packageTitle{TestRcppArmadillo}
}
\description{
\packageDescription{TestRcppArmadillo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TestRcppArmadillo}
\packageIndices{TestRcppArmadillo}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TestRcppArmadillo}

Maintainer: \packageMaintainer{TestRcppArmadillo}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
